requirements.txt
langchain
zenoml>=0.4.0
faiss-cpu
openai